connect api and refactoring
