+++ front-end/src/components/Footer/Footer.tsx
@@ -1,0 +1,9 @@
+const Footer = () => {
+  return (
+    <div className="mx-auto max-w-6xl px-4 pb-8 text-xs text-muted-foreground">
+      Coming soon.
+    </div>
+  );
+};
+
+export default Footer;
